1.Move styles to separate dictionary files. 2.Changed focus visual corner radias to 5. 3.Highlight color had been changed to two separated ones.

+++ Checkem/CustomComponents/AnimatedTextBox.xaml.cs
@@ -15,8 +15,8 @@ namespace Checkem.CustomComponents
         }
 
 
-        private int _textboxFontSize;
-        public int TextboxFontSize
+        private double _textboxFontSize;
+        public double TextboxFontSize
         {
             get
             {
